website: "migrate" to lbssg.
The untitled website generator was renamed to lbssg and it was also
moved within the realm of Libreboot.

Originally this patch was made to address only that, however recently,
the https://notabug.org forge became unavailable. Its webpage now
says:

    NotABug.org

    Due to relentless AI scrapers notabug.org is currently down.

    We are attempting to come up with a solution to this problem.

    We will at a minimum attempt to make repositories available in a
    read-only mode. When exactly this will happen is not clear at the
    moment.

For GNU Boot this creates the following build failure:

    $ make release
    [...]
    ./build.sh
    Cloning into 'untitled'...
    fatal: unable to access 'https://notabug.org/untitled/untitled.git/':
    The requested URL returned error: 504

So without that fix we can't build the website anymore, and we can't
do releases either because resources/packages/website/release also
builds the website.

Only the commit message was modified due to the https://notabug.org
outage.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>