website: pages: remove custom anchors.
The {#<some-anchor-name>} is not defined in the CommonMark standard,
and in it, there is currently no way to add an anchror without using
HTML.

So the way to go for the Haunt migration is to write code that
automatically generates anchor from title names.

Given the status of the documentation on the GNU Boot website, keeping
exactly the same ancor names for these custom anchor is not very
important.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>