coreboot: blobs.list: fam15h: remove bdk-if-phy-vetesse-8514.c.
The init_script_rev_a variable is an array of hexadecimal values. At
first sight it looks like values that are sent to some PHY or Ethernet
controller registers. However the variable name contains init_script
and rev_a. So in doubt it is better to remove it as for now we are
missing the corresponding source code or meaningful names or
documentation that could clarify what this code does.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>