packages: ich9utils: download: workaround unreachable repository.
The notabug.org forge has been down for more than one day, and we
found no place information on the website status and/or outage is
supposed to be given, so we have no idea when or if the forge will be
up again.

Right now we can see a web page with "NotABug is experiencing an
issue." and it encourages to contact the people behind notabug through
mail.

Since we're probably not the only ones depending on notabug they most
likely got tons of mails already so all we can do is wait.

The problem is that we also urgently need to re-do past GNU Boot
source releases to remove nonfree software, so we can't wait too much.

Since the ich9utils git has only one commit and that the code seems
abandonned, it's also safe to host it ourselves.

Since we're in a hurry we hosted it on the forge of an association
that already supports and use GNU Boot, as opening a new repository on
savannah will most likely take time.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>