packages: i945-thinkpads-install-utilities: remove duplicate GNU Boot source release.
In the commit fcccaee8a568259fc6df3484e0dcb361ac62fc33:("packages:
Also release GNU Boot source code."), the GNU Boot source code is
released as a git bundle.

The source code can be retrieved in this way:
    $ mirror="https://ftp.gnu.org/gnu/gnuboot/gnuboot-0.1-rc6/"
    $ wget "${mirror}/gnuboot-source-0.1-rc6.bundle"
    $ wget "${mirror}/gnuboot-source-0.1-rc6.bundle.sig"
    $ gpg --verify gnuboot-source-0.1-rc6.bundle.sig
    $ git clone gnuboot-source-0.1-rc6.bundle gnuboot

So we don't really need a duplicated tarball, especially when we
advise in the official build instructions (in
pages/docs/build/index.md) that "At the time of writing users wanting
to build binaries need to download the GNU Boot source code with git."
because building from tarballs are unsupported and untested.

Building with git bundles is also untested, and it lacks tags, but it
has more probability to work.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>