website: docs: build: handle fragile bin/ and bin-dbg/ handling.
Runnign 'make release', runs './build release roms' in the relase
target in Makefile.am which then runs resources/packages/roms/release
which contains this code:

    if [ ! -d "bin/" ]; then
        ./build boot roms all
    fi

This calls resources/packages/roms/boot which contains this code:

    ./build boot roms_helper "${board}"

This then call resources/packages/roms/boot for each board (with the
board name as argument), which contains:

    [ -d "${romdir}/" ] || mkdir -p "${romdir}/"
    rm -f "${romdir}"/* "${dbgdir}"/*

The dbgdir is then created later on in the same file.

This means that if the build is interrupted in any ways (for instance
if a repository didn't fetch) and is restarted, the build system will
consider the build done.

This can sometimes result in the following error, which is more a
symptom than a problem:

    Building ROM debug image archives for version <version>
    resources/packages/roms/release: line 97: cd: bin-dbg/:
    No such file or directory

In addition just running './build boot roms qemu-pc_2mb' before
running 'make release' will silently fail because qemu-pc_2mb has
debug images, and so no other images will be built, and the release
will only package images for one computer.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>