website: events: 6/7 December 2025: do a separate announce.
The files are meant to submit events to the Agenda du
Libre(https://www.agendadulibre.org).

The result can be previewed with the following command (adl-submit is
available in Guix):

    adl-submit --test-output test.html --file FILE

Note that this command requires network access (it sends information
to the main Agenda du Libre instance). The command above could also
not be documented in the file itself as it contains -- which need to
be escaped even if they are in xml comments.

While this event is already announced as part of the bigger event, we
chose to also announce it separately.

This will make the event a bit more visible and maybe attract more
people than last time, especially people interested in 100% free
software.

This is inspired by an association that tries to address gender bias
on the French Wikipedia that did something similar[1].

[1]https://www.agendadulibre.org/events/33095

As for the supported distributions, we didn't list ProteanOS because
its website is down. Ututo S was also not listed because while it is
probably still active, we the last downloadable medias are from 2017,
and there are other 100% free alternatives that are more up to date.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>