artwork: reduce images size.
Since the commit 82ca3a747cc35e12f852588e5f08164ed8777a8a ("Add GNU
Boot logos sources and use them to generate the GRUB backgrounds."),
we have the following build error:

    Built gigabyte/ga-g41m-es2l (GA-G41M-ES2L)
    E: Could not add [background1280x800.png, 52316 bytes (51 KB)@0x0]; too big?
    E: Failed to add 'background1280x800.png' into ROM image.
    E: Failed while operating on 'COREBOOT' region!
    E: The image will be left unmodified.

Before that commit, the backgrounds were 9572 bytes once included in
the image. After they were 51KB as shown above. This reduces the image
size to 20034 bytes, which is still too much.

The diffoscope program from Trisquel was used to find out the original
PNG settings:
- A palette was used.
- The image data was encoded as hex instead of base64.
- The number of colors is way higher now than it was before the
  82ca3a747cc35e12f852588e5f08164ed8777a8a commit.

  We have 252 colors:
    $ gm identify -format "%k" background1280x800.png
    252
  And before, with the same command, we had 17.

Once compressed by cbfstool with lzma, it goes down to 18421, but it
is still too much. This is because we have only 16996 bytes available:

    $ cbfstool bin/ga-g41m-es2l/seabios_ga-g41m-es2l_txtmode_[KBD].rom print
    [...]
    (empty)                        0xe6040    null            16996 none
    [...]

But if we then compress the GRUB configuration files and we manage to
fit everything. So that will do for now.

In the longer run, more in depth size reduction effort need to be
done, and there is a lot of room for improvement everywhere:
backgrounds size can be reduced, removed from text mode images, GRUB
is compressed twice, has too many modules, etc.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>